add a pause when an unseen before monster appears next to the hero.
When it happens, print a notification message on the screen, and
wait for the player to press space.
Also, reformat some hard-to-read conditional statements in misc.c .
Also, add comments to the modified lines in the source code.

Modified files:
rogue.h: add member t_seen to struct "thing" in rogue.h . It's used to alarm
	the player when an unseen monster suddenly appears next to the hero.
monster.c: In new_monster(): set t_seen to false on the creation of a monster.
misc.c: in look(): pause until a player presses space if a monster appears
	next to the hero. Mark the monster as seen.
room.c: in enter_room(): set t_seen for visible monsters.